requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
WTForms==3.0.1
cx-Oracle==8.3.0
PyMySQL==1.1.0
mysql-connector-python==8.2.0
cryptography==41.0.7
python-dotenv==1.0.0
bcrypt==4.0.1
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==10.0.1
python-magic==0.4.27
webauthn==1.11.1
redis==5.0.1
celery==5.3.4
requests==2.31.0
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
nltk==3.8.1
python-jose==3.3.0
